Compact kafkaexporter files
