--- exporter/kafkaexporter/config.go
@@ -1,11 +1,0 @@
-package kafkaexporter
-
-import (
-	"github.com/IBM/sarama"
-)
-
-type Settings struct {
-	Topic     string
-	Addresses []string
-	*sarama.Config
-}
