--- exporter/kafkaexporter/message_sender.go
@@ -1,9 +1,0 @@
-package kafkaexporter
-
-import (
-	"github.com/IBM/sarama"
-)
-
-type MessageSender interface {
-	SendMessages(msgs []*sarama.ProducerMessage) error
-}
